Set width, remove unused imports
