--- frontend/components/TableContainer/DataTable/GraphCell/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./GraphCell";
